requirements.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dill==0.3.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.0
numpy==1.16.2
oauthlib==3.1.0
opencv-python==3.4.0.14
opt-einsum==3.3.0
Pillow==7.2.0
promise==2.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
tb-nightly==2.4.0a20200916
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-datasets==3.2.1
tensorflow-estimator==2.3.0
tensorflow-metadata==0.23.0
termcolor==1.1.0
tf-estimator-nightly==2.4.0.dev2020091601
tf-nightly==2.4.0.dev20200916
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
